Add scroll QScrollArea to allow use on smaller monitors.
In some cases the list of v4l control is so big that it can't fit on the
monitor any more. For example the Logitech c920. This change adds a
scroll area so that we can scroll through the controls when the window
is smaller than the amount of controls.